* Atleast it's a start to optimize depotsaves... perhaps we should store depotitems as BLOB too?
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1372 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ iologindata.cpp
@@ -900,19 +900,22 @@ bool IOLoginData::savePlayer(Player* player, bool preSave/* = true*/)
 	if(!saveItems(player, itemList, query_insert))
 		return false;
 
-	//save depot items
-	query.str("");
-	query << "DELETE FROM `player_depotitems` WHERE `player_id` = " << player->getGUID();
-	if(!db->executeQuery(query.str()))
-		return false;
+	if(player->depotChange)
+	{
+		//save depot items
+		query.str("");
+		query << "DELETE FROM `player_depotitems` WHERE `player_id` = " << player->getGUID();
+		if(!db->executeQuery(query.str()))
+			return false;
 
-	itemList.clear();
-	for(DepotMap::iterator it = player->depots.begin(); it != player->depots.end(); ++it)
-		itemList.push_back(itemBlock(it->first, it->second));
+		itemList.clear();
+		for(DepotMap::iterator it = player->depots.begin(); it != player->depots.end(); ++it)
+			itemList.push_back(itemBlock(it->first, it->second));
 
-	query_insert.setQuery("INSERT INTO `player_depotitems` (`player_id`, `pid`, `sid`, `itemtype`, `count`, `attributes`) VALUES ");
-	if(!saveItems(player, itemList, query_insert))
-		return false;
+		query_insert.setQuery("INSERT INTO `player_depotitems` (`player_id`, `pid`, `sid`, `itemtype`, `count`, `attributes`) VALUES ");
+		if(!saveItems(player, itemList, query_insert))
+			return false;
+	}
 
 	query.str("");
 	query << "DELETE FROM `player_storage` WHERE `player_id` = " << player->getGUID();
